+++ Gemfile
@@ -5,3 +5,8 @@ gem 'narray'
 group :test do
   gem 'minitest'
 end
+
+group :development do
+  gem 'guard'
+  gem 'guard-minitest'
+end
