+++ lib/mm/dist_config.rb
