+++ lib/mm/helpers.rb
