--- spec/mm_spec.rb
@@ -1,32 +1,0 @@
-require 'minitest/autorun'
-require '../mm.rb'
-
-describe MM do
-  describe "::vector_delta" do
-    describe "with default arguments" do
-      it "should require a vector argument" do
-        
-      end
-    end
-  end
-  
-  describe "::interval_class" do
-    
-  end
-  
-  describe "::num_pairs" do
-    
-  end
-  
-  describe "::sgn" do
-    
-  end
-  
-  describe "::sgn_single" do
-    
-  end
-  
-  describe "::ordered_2_combinations" do
-    
-  end
-end
