+++ spec/spec_helper.rb
@@ -1,1 +1,1 @@
-$: << File.expand_path(File.dirname(__FILE__) + "/../lib")
+$: << File.expand_path(File.dirname(__FILE__) + "/../../lib")
